+++ apps/chat/src/store/marketplace/marketplace.selectors.ts
@@ -97,9 +97,3 @@ export const selectSourceTypes = createSelector(
 
 export const selectDeleteModel = (state: RootState) =>
   rootSelector(state).deleteModel;
-
-export const selectPublishModel = (state: RootState) =>
-  rootSelector(state).publishModel;
-
-export const selectLogsEntity = (state: RootState) =>
-  rootSelector(state).logsEntity;
